refactor: seperate app and server boot
