+++ src/index.js
@@ -1,27 +1,7 @@
-// content of index.js
-const express = require("express");
-const port = process.env.PORT || 6060;
-const app = express();
+const app = require('./app');
 const db = require("./config/db");
-const bodyParser = require("body-parser");
-const authRoutes = require("./modules/auth/routes");
-const quoteRoutes = require("./modules/quotes/routes");
 
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use((req, res, next) => {
-  res.append("Access-Control-Allow-Origin", ["*"]);
-  res.append("Access-Control-Allow-Methods", "GET,PUT,POST,DELETE");
-  res.append("Access-Control-Allow-Headers", "Content-Type");
-  next();
-});
-
-require("babel-register")({
-  presets: ["env"]
-});
-
-app.use("/auth/", authRoutes);
-app.use("/quotes/", quoteRoutes);
+const port = process.env.PORT || 6060;
 
 app.listen(port, err => {
   if (err) {
@@ -30,5 +10,3 @@ app.listen(port, err => {
   db();
   console.log(`server is listening on ${port}`);
 });
-
-module.exports = app;
